+++ src/config.c
@@ -210,9 +210,9 @@ const config default_cfg = {
   exit_write : false,
   exit_title : W(""),
   icon : W(""),
-  log : W(""),
-  logging : true,
-  title :  W(""),
+  log : W("fatty.$h.%Y-%m-%d_%H-%M-%S.$p.log"),
+  logging : false,
+  title : W(""),
   create_utmp : false,
   window : 0,
   x : 0,
@@ -1235,7 +1235,7 @@ init_config_dirs(void)
   if (config_dirs)
     return;
 
-  int ncd = 3;
+  int ncd = 4;
   char * appdata = getenv("APPDATA");
   if (appdata)
     ncd++;
@@ -1244,6 +1244,7 @@ init_config_dirs(void)
   config_dirs = newn(string, ncd);
 
   // /usr/share/mintty , $APPDATA/mintty , ~/.config/mintty , ~/.mintty
+  config_dirs[++last_config_dir] = "/usr/share";  // for "/emojis" only
   config_dirs[++last_config_dir] = "/usr/share/fatty";
   if (appdata) {
     appdata = newn(char, strlen(appdata) + 8);
@@ -2084,7 +2085,10 @@ do_file_resources(control *ctrl, wstring pattern, bool list_dirs, str_fn fnh)
   init_config_dirs();
   //printf("add_file_resources <%ls> dirs %d\n", pattern, list_dirs);
 
-  for (int i = last_config_dir; i >= 0; i--) {
+  int base = 1;
+  if (0 == wcsncmp(W("emojis/"), pattern, 7))
+    base = 0;  // also look up distinct emojis packages in /usr/share/emojis
+  for (int i = last_config_dir; i >= base; i--) {
 #ifdef use_findfile
     (void)fnh;  // CYGWIN_VERSION_API_MINOR < 74